+++ range_cache.h
@@ -506,7 +506,7 @@ protected:
 			uint32_t jumps = 0;
 			if(tops.size() > 0) {
 				entTop = tops.back();
-				jumps = tops.size();
+				jumps = (uint32_t)tops.size();
 			}
 			// Cache the entry for the end of the tunnel
 			assert(map_.find(entTop) == map_.end());
